Remove unreferenced LogEntry structure
The `LogEntryType` enum is used in a number of places, but the `LogEntry`
structure is never used anywhere.  I remember this causing me no end of
confusion when I was grappling with understanding RFC6962, and since the
structure hasn't seen any new use since, I'm thinking it can probably stand
to go.