Remove X509ChainEntry, clarify content of extra_data
After `LogEntry` got nuked, I realised that `X509ChainEntry` was now also
unused as a structure.  So, I nuked that.  It was referenced in a couple of
places in the *text*, so I fixed those references up so they're now more
clear as to what they're *actually* talking about.  Since I needed something
to represent a certificate chain, I created `CertificateChain` as a
top-level structure.  I also cleaned up the structure definitions in a
couple of places:

* Once we've got `CertificateChain` defined, it's pretty clear that that's
  what the second element in PrecertChainEntryV2 is; and
* I didn't see any reason to have a case statement to choose between a
  `TBSCertificate` or a `TBSCertificate`, so I just made `signed_entry` in
  `SignedCertificateTimestamp` a straight-up member.

For anyone playing along at home, the definition of `extra_data` in
`/get-entries` is what tripped up alpha.ctlogs.org so badly, so I'm glad to
be able to clarify this for the future.