Correct versioning of some calls.
The example using v1 get-entries is bad as a v1 log is unlikely to return the
error codes described in bis.

add-pre-chain requires a version upgrade as it's specified to return the same
outputs as add-chain, which now returns the SCT in binary format.

get-sth may need updating as well since entries are structured differently.