Bashing more client behaviour statements.
The RFC should specify how to determine whether a certificate is CT-compliant
or not, by specifying the requirement for SCTs to be present alongside it
during the TLS handshake and how to check their validity.
Once the compliance status of the certificate has been established, it is
up to the client to decide how it behaves.

This commit attempts to reflect that.