Ticket 156: Requiring SCTs in addition to, not instead of, inclusion proofs.
The gist of the reasoning is that without the timestamp and extensions fields
from the SCT, it is not possible to calculate the leaf hash that would
be used together with the rest of the nodes from the inclusion proof to
calculate the root hash.

To make implementation of clients simpler (i.e. clients would only need
to check signatures from SCTs rather than having to check them over STHs
as well and implement inclusion checking) this PR suggests mandating
presence of SCTs, while pointing out the benefit of supplying both
SCT and inclusion proof.