Allow a log to reject badly-encoded submissions.
However if the log chooses to accept a badly-encoded submission it must
issue an SCT.

Also, remove the sentence about lack of ability to detect encoding errors.

This sentence is completely accurate but, as far as I can tell, does not
contribute to the understanding of the expected log's behaviour in this
case beyond what's stated above that logs may log certs not compliant
with RFC5280.