Section 3.1: Clarify the "may" is not permissive. (#168)
* Section 3.1: Clarify the "may" is not permissive.

"may" sounds permissive, "might" sounds better; otherwise this sounds like a permission and not a prediction.
Alternatively, the proposed text...

* Updating with Rob's suggestion.

This sounds better than my original proposed. 

My original concern was that people reading RFCs often have learnt to scan for MUST/must, MAY/may and implicitly try interpret them in RFC2119 context. If we want to make this the document that people reference when justifying not accepting logged cert, that's fine, but then we should be more explicit (and I don't think that this was the original intent of the sentence). 
This is also just a nit, the PR can be closed, ignored, giggled at, etc.