Specify policy for IANA registries.
Specify the policy for adding values to the various IANA registries
required in 6962-bis.
An expert review is required for the signature and hash algorithms, as
new values do not affect the protocol overall, but certain properties
of these algorithms still have to be vetted by an expect.

A specification is required for the STH and SCT extension registries
as well as the VersionedTransType registry.
An expert should review the specification for interoperability.

An expert review is required for requesting OIDs from the first
Log ID repository, as it contains a limited number of OIDs.

OIDs from the second Log ID repository are handed on a "First Come
First Serve" basis.

All policies are defined in RFC5226, so a reference to it has been added.