Issue #181 - Consolidate 'add-chain' and 'add-pre-chain'
This is a different take on bifurication's pull request (PR #238):
* The certificate submitted has to be explicitly specified in
a separate parameter, rather than implied from the presence of the
precertificate parameter.
* An error code has been added to indicate if a submission is invalid
because both parameters have been specified.

I chose this approach as it explicitly indicates what's the certificate
the submitter wishes to be included in the log, avoiding the potential
issue of logging partial chains because the precertificate was omitted.