Issue 167: Defining incorporation
Define what it means to incorporate a submission into the tree.

Additionally, clarify what a log MUST do to adhere with this
requirement:
Rather than dictate that this is done periodically, state it MUST
be done within the MMD for each submission, pointing out that batching
of the operation as an implementation detail.