A-D review: Re-addressing of the MUSTs->SHOUDLs.
Elaborate on why, despite not having any security implications, it is
not wise to allow log implementations to admit submissions that do not
chain to a trusted root anchor.

Note the phrase "with a valid signature chain" was omitted, as the
following paragraph deals with it.

It should be clear that the MUST in this paragraph is to guide
implementations to never admit submissions that do not chain to an
accepted trust anchor - while implementations may have bugs with that
result, the option to admit such submissions is not at the
implementation's discretion (using a SHOULD would imply that it is,
under certain circumstances. a MUST clearly indicates it is not).

Hopefully this clarification and the removal of the wording about valid
signature chains makes it clear that the MUST is intended to guide
implementations in case of submissions not chaining to an accepted trust
anchor.